backend/requirements.txt
# Data Processing
pandas==2.1.3
numpy==1.26.2
requests==2.31.0
beautifulsoup4==4.12.2
PyYAML==6.0.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# API
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.4.2
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Testing
pytest==7.4.3
httpx==0.25.1
pytest-cov==4.1.0

# Utilities
tqdm==4.66.1
loguru==0.7.2 